Separate dynamichmc part to its own extension
