+++ Project.toml
@@ -32,10 +32,12 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 [weakdeps]
 MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
 SampleChains = "754583d1-7fc4-4dab-93b5-5eaca5c9622e"
+SampleChainsDynamicHMC = "6d9fd711-e8b2-4778-9c70-c1dfb499d4c4" 
 
 [extensions]
 ArviZMCMCChainsExt = "MCMCChains"
 ArviZSampleChainsExt = "SampleChains"
+ArviZSampleChainsDynamicHMCExt = ["SampleChains", "SampleChainsDynamicHMC"]
 
 [compat]
 DataInterpolations = "4"
