+++ src/ArviZ.jl
@@ -89,6 +89,9 @@ include("conversions.jl")
         @require SampleChains = "754583d1-7fc4-4dab-93b5-5eaca5c9622e" begin
             include("../ext/ArviZSampleChainsExt.jl")
         end
+        @require SampleChainsDynamicHMC = "6d9fd711-e8b2-4778-9c70-c1dfb499d4c4" begin
+            include("../ext/ArviZSampleChainsDynamicHMCExt.jl")
+        end
         @require MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d" begin
             include("../ext/ArviZMCMCChainsExt.jl")
         end
